Try to prevent notifications immediately after alarm creation
I have seen the occasional OK message being posted to Slack after an instance has launched. After checking the Lambda event and alarm history, it looks like it should have detected that it was the first state update and not posted to Slack. I suspect that it is a problem of eventual consistency and describe_alarm_history returned 0 state updates at the time, while this specifically checking for 1 state update. This changes it so that when there are 0 or 1 state updates, it is deemed as the first state update and it gets ignored.